fix profiles, add ad hoc localization
to localize menu in russian you have to rename unpager.exe to
unpager_ru.exe

+++ unpager/Polynomial.cs
@@ -57,17 +57,17 @@ namespace WindowsFormsApplication1
                     A[i, j] = 0.0;
                     foreach (Point p in carcas)
                     {
-                        A[i, j] += Scalar.pow(p.X, i + j);
+                        A[i, j] += Scalar.pow((double)p.X, i + j);
                     }
                     foreach (Point p in firm_carcas)
                     {
-                        A[i, j] += Scalar.pow(p.X, i + j) * FIRM_MULTIPLIER;
+                        A[i, j] += Scalar.pow((double)p.X, i + j) * FIRM_MULTIPLIER;
                     }
                 }
                 B[i] = 0.0;
                 foreach (Point p in carcas)
                 {
-                    B[i] += p.Y * Scalar.pow(p.X, i);
+                    B[i] += p.Y * Scalar.pow((double)p.X, i);
                 }
                 foreach (Point p in firm_carcas)
                 {
